@@ -18,6 +18,7 @@
     "pack:main": "cross-env NODE_ENV=production webpack --progress --colors --config .electron-vue/webpack.main.config.js",
     "pack:renderer": "cross-env NODE_ENV=production webpack --progress --colors --config .electron-vue/webpack.renderer.config.js",
     "postinstall": "",
+    "start":"nodemon --watch src/main --watch src/shared --exec './node_modules/.bin/electron' ./out/main/main.js",
     "docs": "docsify serve docs"
   },
   "build": {
@@ -99,8 +100,8 @@
     "vue-loader": "^15.2.4",
     "vue-style-loader": "^4.1.0",
     "vue-template-compiler": "^2.6.7",
-    "webpack": "^4.15.1",
-    "webpack-cli": "^3.0.8",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0",
     "webpack-hot-middleware": "^2.22.2"
   },
@@ -110,7 +111,7 @@
     "cheerio": "^1.0.0-rc.2",
     "custom-electron-titlebar": "^3.2.2-hotfix62",
     "electron-log": "^3.0.1",
-    "electron-updater": "^4.0.6",
+    "electron-updater": "^4.3.1",
     "element-ui": "^2.5.4",
     "fs": "^0.0.1-security",
     "http2-client": "^1.3.3",
@@ -119,7 +120,7 @@
     "nodemailer": "^5.1.1",
     "path": "^0.12.7",
     "qs": "^6.7.0",
-    "request": "^2.88.0",
+    "request": "^2.88.2",
     "vue": "^2.6.10",
     "vue-electron": "^1.0.6",
     "vue-router": "^3.0.1",
