@@ -119,13 +119,15 @@ function startElectron () {
   ]
 
   // detect yarn or npm and process commandline args accordingly
-  if (process.env.npm_execpath.endsWith('yarn.js')) {
-    args = args.concat(process.argv.slice(3))
-  } else if (process.env.npm_execpath.endsWith('npm-cli.js')) {
-    args = args.concat(process.argv.slice(2))
-  }
+  // if (process.env.npm_execpath.endsWith('yarn.js')) {
+  //   args = args.concat(process.argv.slice(3))
+  // } else if (process.env.npm_execpath.endsWith('npm-cli.js')) {
+  //   args = args.concat(process.argv.slice(2))
+  // }
 
-  electronProcess = spawn(electron, args)
+  args = args.concat(process.argv.slice(2))
+
+  electronProcess = spawn(electron, args,{shell: true})
   
   electronProcess.stdout.on('data', data => {
     electronLog(data, 'blue')
@@ -180,7 +182,7 @@ function init () {
   Promise.all([startRenderer(), startMain()])
     .then(() => {
       console.log('==========22222')
-      startElectron()
+      // startElectron()
     })
     .catch(err => {
       console.error(err)
