@@ -1,23 +1,25 @@
-const elog = require('electron-log')
 
-elog.transports.file.level = 'debug'
+import elog from 'electron-log'
+import os from 'os'
 
+export { version } from '../../package'
+
+export const isDev = process.env.NODE_ENV === 'development'
+export const isPro = process.env.NODE_ENV === 'production'
 
 /**
  * Set `__static` path to static files in production
  * https://simulatedgreg.gitbooks.io/electron-vue/content/en/using-static-assets.html
  */
-
-if (process.env.NODE_ENV !== 'development') {
-    // eslint-disable-next-line no-underscore-dangle
+if (isPro) {
     global.__static = require('path').join(__dirname, '/static').replace(/\\/g, '\\\\')
 }
 
+
+elog.transports.file.level = 'debug'
 export const log = elog
 
-const os = require('os')
-
-export const baseUrl = process.env.NODE_ENV === 'development' ? 'http://localhost:9080' : `file://${__dirname}/index.html`
+export const baseUrl = isDev ? 'http://localhost:9080' : `file://${__dirname}/index.html`
 
 const OSTYPES = {
     Darwin: 'mac',
@@ -25,17 +27,8 @@ const OSTYPES = {
 }
 
 export const osType = OSTYPES[os.type()] || 'win'
-export function isMac() {
-    return osType === 'mac'
-}
-
-export function isWin() {
-    return osType === 'win'
-}
-
-export function isDev() {
-    return process.env.NODE_ENV === 'development'
-}
+export const isMac = osType === 'mac'
+export const isWin = osType === 'win'
 
 export const imageSourceType = [
     {
